update: 完善播放器解码 fix: stream播放器名称 feat: 可以在调用Play方法时选择是否阻塞当前线程，以便于等待音频播放完毕后继续执行后面的程序

+++ src/server/XiaoZhi.Net.Server.AudioPlayer.Abstractions/IAudioPlayer.cs
@@ -21,7 +21,7 @@ public interface IAudioPlayer : IDisposable
     /// <summary>
     /// Event that is raised when audio data is available.
     /// </summary>
-    event Action<byte[]> OnAudioDataAvailable;
+    event Action<float[]> OnAudioDataAvailable;
 
     /// <summary>
     /// Gets a value indicating whether FFmpeg has been successfully initialized.
